Fixed shown override control host in the developer settings

+++ app/src/main/java/at/rtr/rmbt/android/ui/fragment/SettingsFragment.kt
@@ -280,7 +280,7 @@ class SettingsFragment : BaseFragment(), InputSettingDialog.Callback, ServerSele
             }
 
             KEY_DEVELOPER_CONTROL_SERVER_PORT_CODE -> {
-                settingsViewModel.state.controlServerPort.set(value.toInt())
+                settingsViewModel.state.controlServerPort.set(value)
             }
             KEY_DEVELOPER_MAP_SERVER_HOST_CODE -> {
                 settingsViewModel.state.mapServerHost.set(value)
